Combine benchmarks to speed up build time
