+++ Cargo.toml
@@ -146,13 +146,5 @@ path = "src/main.rs"
 # Run all benchmarks with `cargo bench`
 # Run individual benchmarks like `cargo bench -- <regex>` e.g. `cargo bench -- parse`
 [[bench]]
-name = "encoder_benchmark"
-harness = false
-
-[[bench]]
-name = "eval_benchmark"
-harness = false
-
-[[bench]]
-name = "parser_benchmark"
+name = "benchmarks"
 harness = false
